requirements.txt
matplotlib
numpy>=1.20
roboticstoolbox-python
scikit-learn
